requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiocontextvars==0.2.2     # via loguru
attrs==19.3.0             # via pytest
autopep8==1.5
click==7.0                # via flask
contextvars==2.4          # via aiocontextvars
coverage==5.0.3           # via pytest-cov
cycler==0.10.0            # via matplotlib
dataclasses==0.7          # via screeninfo
flask==1.1.1
immutables==0.11          # via contextvars
importlib-metadata==1.5.0  # via pluggy, pytest
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via flask
joblib==0.14.1            # via scikit-learn
kiwisolver==1.1.0         # via matplotlib
loguru==0.4.1
markupsafe==1.1.1         # via jinja2
matplotlib==3.1.3
more-itertools==8.2.0     # via pytest
numpy==1.18.1
opencv-python==4.2.0.32
packaging==20.1           # via pytest
pandas==1.0.1
pillow==7.0.0
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via autopep8
pyparsing==2.4.6          # via matplotlib, packaging
pytest-cov==2.8.1
pytest==5.3.2
python-dateutil==2.8.1    # via matplotlib, pandas
pytz==2019.3              # via pandas
pyyaml==5.3
scikit-learn==0.22.1      # via sklearn
scipy==1.4.1
screeninfo==0.6.1
six==1.14.0               # via cycler, packaging, python-dateutil
sklearn==0.0
wcwidth==0.1.8            # via pytest
werkzeug==1.0.0           # via flask
zipp==3.0.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
